Sequential Dependence Model
An implementation of the sequential dependence part of (A Markov Random
Field Model for Term
Dependencies)[https://ciir.cs.umass.edu/pubfiles/ir-387.pdf]. See #30.

There are a number of improvements and optimizations that need to be
addressed which can be found in #30.

There are also a number of improvements to the index. Therefore this PR
is landing in master to facilitate the building of the new index
structures before this feature is enabled for extraction.

Notes regarding changes:

* Marked `calculate_lm` inline. This avoids multiple definition
  compile errors when running the test suite. This will be fixed by #34.

* Follows Indri's implementation:

    - indri::infnet::ContextCountAccumulator
    - indri::infnet::UnorderedWindowNode
    - indri::infnet::OrderedWindowNode
    - indri::infnet::WeightedAndNode

* There is no support for running queries over fields within a document.

* Add `DirichletTermScore` a port of Indri's QL scoring function. Some
  features were omitted because they are rarely used (for example
  document smoothing with the `docMu` parameter).

* Test with query 'image segway example' exercises duplicate checking
  pathway in `count_unordered_phrase`.

* Encapsulate all FastPFor operations into a single implementation file
  (compression.cpp). This will make link time operations easier,
  particularly for the test suite.

* Make `Document` id aware

* Add `dump_fixture` program for generated test indexes.

* Add `length` method to `query_train_file`.