Intersect postings to score SDM bigrams
This patch improves the efficiency of computing bigram statistics (both
ordered and unorderd) when scoring documents using `Sdm`.

A scoring context is setup when `Sdm::extract` is called. In
`Sdm::set_context` the relevant data structures are created so that
statistics are computed for only the documents that inlcude occurrences
of both terms in the bigram.

Part of #30