Fix postings organization during indexing
When building the `InvertedIndex`, the `IndexerInteractor` was
storing posting entries in the wrong 'slots' (`InvertedIndex` is a
`std::vector` where the postings are placed at the index of the term id
for that postings list).

The `IndexerInteractor` first builds the `Lexicon`, and then later it
builds the `InvertedIndex`. Both of these data structures rely on the
term id's from Indri.

When building the `Lexicon`, Indri's `indri::index::VocabularyIterator`
is used which enumerates Indri's vocabulary in ascending term id order.
However, later when building the `InvertedIndex`, Indri's
`indri::index::DocListFileIterator` is used and traverses the postings
(and therefore term id's) in a different order (the ordering may be
related to frequent and infrequent term data structures within Indri,
this is not confirmed, the point is that it *not* ascending term id's).

An existing solution to this problem was implemented incorrectly. It
used a vector `idx_lexid_` to track (term id, position) information
during `IndexerInteractor::lexicon` and then use `idx_lexid_` later in
`IndexerInteractor::inverted_index` to put the postings in the same
'slot'. This method completely ignored the 'current' term id that was
being processed during `InvertedIndex` construction, causing postings to
be placed in the wrong 'slots' in the `InvertedIndex` vector.

The fix is simple. Fetch the term id to place the postings in the
correct 'slot' in the `InvertedIndex` vector. Also a sanity check is
performed during `Lexicon` construction to confirm that the term id's
are being enumerated in ascending order of term id.

See #48