New header file for web document features
Replaced by the old `wikipedia` header to make it easier to test all
static document features.

This commit renames the old `wikipedia` headers and test files to a more
general `web_document` and moves the static document feature
`set_url_lendep` into the new header to prepare for a testing and a bug
fix.